refactor(keys): create keys directory
