+++ src/SubtleCrypto.js
@@ -1,11 +1,11 @@
 /**
  * Local dependencies
  */
-const CryptoKey = require('./CryptoKey')
-const CryptoKeyPair = require('./CryptoKeyPair')
-const JsonWebKey = require('./JsonWebKey')
+const CryptoKey = require('./keys/CryptoKey')
+const CryptoKeyPair = require('./keys/CryptoKeyPair')
+const JsonWebKey = require('./keys/JsonWebKey')
+const recognizedKeyUsages = require('./keys/recognizedKeyUsages')
 const supportedAlgorithms = require('./algorithms')
-const recognizedKeyUsages = require('./recognizedKeyUsages')
 const InvalidAccessError = require('./errors/InvalidAccessError')
 const NotSupportedError = require('./errors/NotSupportedError')
