+++ src/dictionaries/RsaHashedKeyAlgorithm.js
@@ -5,14 +5,14 @@ const RSA = require('node-rsa')
 const crypto = require('crypto')
 const {spawnSync} = require('child_process')
 const {pem2jwk, jwk2pem} = require('pem-jwk')
+const {TextEncoder, TextDecoder} = require('text-encoding')
 
 /**
  * Local dependencies
  */
-const {TextEncoder, TextDecoder} = require('text-encoding')
-const CryptoKey = require('../CryptoKey')
-const CryptoKeyPair = require('../CryptoKeyPair')
-const JsonWebKey = require('../JsonWebKey')
+const CryptoKey = require('../keys/CryptoKey')
+const CryptoKeyPair = require('../keys/CryptoKeyPair')
+const JsonWebKey = require('../keys/JsonWebKey')
 const KeyAlgorithm = require('./KeyAlgorithm')
 const RsaKeyAlgorithm = require('./RsaKeyAlgorithm')
 const supportedAlgorithms = require('../algorithms')
